src/test/java/nextstep/refactor_ladder/domain/LineTest.java
package nextstep.refactor_ladder.domain;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.util.Collections;

import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
import static org.junit.jupiter.api.Assertions.assertAll;

class LineTest {
    @Test
    @DisplayName("정보가 빈값이거나 null인 경우 예외 발생 검증")
    void create_exception() {
        assertAll(() -> {
            assertThatIllegalArgumentException().isThrownBy(() -> Line.from(Collections.emptyList()));
            assertThatIllegalArgumentException().isThrownBy(() -> Line.from(null));
        });
    }
}
